// client/src/entities/RepairOrder.ts
type RepairOrder = {
  id: number;
  productName: string;
  description: string;
  details: string;
  cost: string;
  status: string;
  estimateFinishDate: string;
  finishDate: string;
  createdAt: string;
  updatedAt: string;
  customerId: number;
  staffId: number;
};
export default RepairOrder;
// "id": 1,
//       "productName": "test product name",
//       "description": "test repair order",
//       "details": "details test repair order",
//       "cost": "100000",
//       "status": "pending",
//       "estimateFinishDate": "2024-12-01",
//       "finishDate": null,
//       "createdAt": "2024-11-24T08:27:17.780Z",
//       "updatedAt": "2024-11-24T08:27:17.780Z",
//       "customerId": 1,
//       "staffId": 1